Time to introduce my new command handler! And many other things...

+++ events/messageReactionAdd.js
@@ -1,9 +1,13 @@
+const nconf = require('nconf');
+
 module.exports = {
   fn: (main, messageReaction, user) => {
     main.paginationHelper.handleMessageReactionAddEvent(messageReaction, user);
 
-    if (messageReaction.emoji.name === '❌' && main.owner.includes(user.id) && messageReaction.message.deletable) {
-      messageReaction.message.delete();
+    if (!nconf.get('bot:selfbot')) {
+      if (messageReaction.emoji.name === '❌' && main.owner.includes(user.id) && messageReaction.message.deletable) {
+        messageReaction.message.delete();
+      }
     }
   },
 };
